requirements.txt
annotated-types==0.7.0
anyio==4.8.0
beautifulsoup4==4.13.3
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
decorator==5.2.1
distro==1.9.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
jiter==0.9.0
joblib==1.4.2
moviepy==2.1.2
mysql-connector-python==9.2.0
nltk==3.9.1
numpy==2.2.3
openai==1.66.3
pillow==10.4.0
proglog==0.1.10
pydantic==2.10.6
pydantic_core==2.27.2
pydub==0.25.1
python-dotenv==1.0.1
python-telegram-bot==21.11.1
regex==2024.11.6
requests==2.32.3
scikit-learn==1.6.1
scipy==1.15.2
sniffio==1.3.1
soupsieve==2.6
threadpoolctl==3.5.0
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.3.0
